+++ crates/hotshot/src/tasks/task_state.rs
@@ -229,7 +229,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> CreateTaskState<TYPES, I>
             spawned_tasks: BTreeMap::new(),
             formed_upgrade_certificate: None,
             proposal_cert: None,
-            version: Arc::clone(&handle.hotshot.version),
             output_event_stream: handle.hotshot.external_event_stream.0.clone(),
             current_proposal: None,
             id: handle.hotshot.id,
@@ -266,7 +265,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> CreateTaskState<TYPES, I>
             output_event_stream: handle.hotshot.external_event_stream.0.clone(),
             id: handle.hotshot.id,
             storage: Arc::clone(&handle.storage),
-            version: Arc::clone(&handle.hotshot.version),
             decided_upgrade_certificate: Arc::clone(&handle.hotshot.decided_upgrade_certificate),
         }
     }
@@ -299,7 +297,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> CreateTaskState<TYPES, I>
             timeout_task,
             round_start_delay: handle.hotshot.config.round_start_delay,
             id: handle.hotshot.id,
-            version: Arc::clone(&handle.hotshot.version),
             formed_upgrade_certificate: None,
             decided_upgrade_certificate: Arc::clone(&handle.hotshot.decided_upgrade_certificate),
         }
@@ -335,7 +332,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> CreateTaskState<TYPES, I>
             spawned_tasks: BTreeMap::new(),
             instance_state: handle.hotshot.instance_state(),
             id: handle.hotshot.id,
-            version: Arc::clone(&handle.hotshot.version),
             decided_upgrade_certificate: Arc::clone(&handle.hotshot.decided_upgrade_certificate),
         }
     }
@@ -369,7 +365,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> CreateTaskState<TYPES, I>
             consensus: OuterConsensus::new(consensus),
             last_decided_view: handle.cur_view().await,
             id: handle.hotshot.id,
-            version: Arc::clone(&handle.hotshot.version),
             decided_upgrade_certificate: Arc::clone(&handle.hotshot.decided_upgrade_certificate),
         }
     }
