+++ crates/task-impls/src/consensus/mod.rs
@@ -31,7 +31,6 @@ use jf_vid::VidScheme;
 #[cfg(async_executor_impl = "tokio")]
 use tokio::task::JoinHandle;
 use tracing::{debug, error, info, instrument, warn};
-use vbs::version::Version;
 
 use crate::{
     consensus::handlers::{
@@ -112,9 +111,6 @@ pub struct ConsensusTaskState<TYPES: NodeType, I: NodeImplementation<TYPES>> {
 
     /// last View Sync Certificate or Timeout Certificate this node formed.
     pub proposal_cert: Option<ViewChangeEvidence<TYPES>>,
-
-    /// Globally shared reference to the current network version.
-    pub version: Arc<RwLock<Version>>,
 
     /// Output events to application
     pub output_event_stream: async_broadcast::Sender<Event<TYPES>>,
@@ -212,7 +208,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
             self.payload_commitment_and_metadata.clone(),
             self.proposal_cert.clone(),
             Arc::clone(&self.instance_state),
-            *self.version.read().await,
             self.id,
         )
         .await?;
@@ -247,7 +242,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
         let quorum_mem = Arc::clone(&self.quorum_membership);
         let da_mem = Arc::clone(&self.da_membership);
         let instance_state = Arc::clone(&self.instance_state);
-        let version = *self.version.read().await;
         let id = self.id;
         let handle = async_spawn(async move {
             update_state_and_vote_if_able::<TYPES, I>(
@@ -259,7 +253,6 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
                 quorum_mem,
                 instance_state,
                 (priv_key, upgrade, da_mem, event_stream),
-                version,
                 id,
             )
             .await;
@@ -274,18 +267,11 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
         event: Arc<HotShotEvent<TYPES>>,
         event_stream: Sender<Arc<HotShotEvent<TYPES>>>,
     ) {
-        let version = *self.version.read().await;
         match event.as_ref() {
             HotShotEvent::QuorumProposalRecv(proposal, sender) => {
                 debug!("proposal recv view: {:?}", proposal.data.view_number());
-                match handle_quorum_proposal_recv(
-                    proposal,
-                    sender,
-                    event_stream.clone(),
-                    self,
-                    version,
-                )
-                .await
+                match handle_quorum_proposal_recv(proposal, sender, event_stream.clone(), self)
+                    .await
                 {
                     Ok(Some(current_proposal)) => {
                         let view = current_proposal.view_number();
@@ -495,22 +481,14 @@ impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ConsensusTaskState<TYPES, I>
 
                 let old_view_number = self.cur_view;
 
-                // If we have a decided upgrade certificate,
-                // we may need to upgrade the protocol version on a view change.
+                // If we have a decided upgrade certificate, the protocol version may also have
+                // been upgraded.
                 if let Some(cert) = self.decided_upgrade_certificate.read().await.clone() {
                     if new_view == cert.data.new_version_first_view {
                         warn!(
-                            "Updating version based on a decided upgrade cert: {:?}",
+                            "Version upgraded based on a decided upgrade cert: {:?}",
                             cert
                         );
-                        let mut version = self.version.write().await;
-                        *version = cert.data.new_version;
-
-                        broadcast_event(
-                            Arc::new(HotShotEvent::VersionUpgrade(cert.data.new_version)),
-                            &event_stream,
-                        )
-                        .await;
                     }
                 }
