+++ crates/task-impls/src/quorum_proposal_recv/mod.rs
@@ -92,9 +92,6 @@ pub struct QuorumProposalRecvTaskState<TYPES: NodeType, I: NodeImplementation<TY
     /// The node's id
     pub id: u64,
 
-    /// Globally shared reference to the current network version.
-    pub version: Arc<RwLock<Version>>,
-
     /// An upgrade certificate that has been decided on, if any.
     pub decided_upgrade_certificate: Arc<RwLock<Option<UpgradeCertificate<TYPES>>>>,
 }
