+++ crates/types/src/message.rs
@@ -18,7 +18,7 @@ use vbs::{
 use crate::{
     data::{DaProposal, Leaf, QuorumProposal, UpgradeProposal, VidDisperseShare},
     simple_certificate::{
-        DaCertificate, UpgradeCertificate, ViewSyncCommitCertificate2,
+        version, DaCertificate, UpgradeCertificate, ViewSyncCommitCertificate2,
         ViewSyncFinalizeCertificate2, ViewSyncPreCommitCertificate2,
     },
     simple_vote::{
@@ -33,34 +33,6 @@ use crate::{
     },
     vote::HasViewNumber,
 };
-
-/// Calculate the version applied in a view, based on the provided upgrade certificate.
-///
-/// # Errors
-/// Returns an error if we do not support the version required by the upgrade certificate.
-pub fn version<TYPES: NodeType>(
-    view: TYPES::Time,
-    upgrade_certificate: &Option<UpgradeCertificate<TYPES>>,
-) -> Result<Version> {
-    let version = match upgrade_certificate {
-        Some(ref cert) => {
-            if view >= cert.data.new_version_first_view
-                && cert.data.new_version == TYPES::Upgrade::VERSION
-            {
-                TYPES::Upgrade::VERSION
-            } else if view >= cert.data.new_version_first_view
-                && cert.data.new_version != TYPES::Upgrade::VERSION
-            {
-                bail!("The network has upgraded to a new version that we do not support!");
-            } else {
-                TYPES::Base::VERSION
-            }
-        }
-        None => TYPES::Base::VERSION,
-    };
-
-    Ok(version)
-}
 
 /// Incoming message
 #[derive(Serialize, Deserialize, Clone, Derivative, PartialEq, Eq, Hash)]
